NXP-24360: isolate test checking probe status count to a dedicated test to avoid collisions with other tests logics
